Fabric/Fabric-CICD/requirements.txt
azure-identity>=1.15.0
fabric-cicd>=0.1.0
PyYAML>=6.0.1
